Move Nav to views
