+++ web/src/views/Nav.vue
@@ -1,9 +1,6 @@
 <template>
   <div id="nav">
     <img alt="Vue logo" src="../assets/logo.png" />
-
-    <router-link to="/">Home</router-link>
-    <router-link to="/about">About</router-link>
   </div>
 
 </template>
